+++ tonic/datasets/pokerdvs.py
@@ -80,7 +80,7 @@ class POKERDVS(VisionDataset):
             events = self.transform(events, self.sensor_size, self.ordering)
         if self.target_transform is not None:
             target = self.target_transform(target)
-        return events.astype("int64"), target
+        return events, target
 
     def __len__(self):
         return len(self.data)
